// src/generated/client/startup_intelligence/research/v1/service_client.ts
export interface ListTechEventsResponse {
  success: boolean;
  count: number;
  conferenceCount: number;
  mappableCount: number;
  lastUpdated: string;
  events: TechEvent[];
  error: string;
}

export interface TechEvent {
  id: string;
  title: string;
  type: string;
  location: string;
  coords?: TechEventCoords;
  startDate: string;
  endDate: string;
  url: string;
  source: string;
  description: string;
}

export interface TechEventCoords {
  lat: number;
  lng: number;
  country: string;
  original: string;
  virtual: boolean;
}